Add check for mutating arguments and switch order
This adds a check for a program which mutates the arguments passed in. It is good practice that you generally shouldn't mutate arguments passed in
without being explicit about it. 

This also fixes some logic for if a code had a newline at the end of their code. It will not longer throw an erroneous centering error for the newline.